Add filter to allow normalisation of term name and slug
As the latest version of the plugin is using `sanitize_term_field` instead of the e `sanitize_title_for_query` the name and slug are no longer being transformed to lowercase (`sanitize_title_for_query` does that while `sanitize_term_field` does not).

This patch is adding a new filter, similar to `es_meta_query_meta_value` one which is dealing with similar issue, in order to allow adapters to fine control the normalisation. As not all adapters are necessary mapping post name or slug to lowecase field, this approach feels better than hardcoding `strtolower` to the plugin itself.